Correctly return stdout for get_distribution_and_version
In the case that the 'lsb_release -d' command fails, the
get_distribution_and_version test should run:
    'cat /etc/*-release | head -1'
using the ssh_client.execute_command. However the result
from this is a dictionary of values. This means this method
returns different types of objects depending on the success
or failure of the call. This causes issues when using this
method.

- Added .stdout call to the return statement

Change-Id: I8cc1840c615e71a3df2ea19c3bda955c81bf7749