Added name_prefix to create server behaviors in compute
Added name_prefix to create_server_with_defaults and
create_active_server. This will allow test users to specify a
name prefix for creating a server. This will be used with
random_name to create a server with a name that contains
the name_prefix value followed by random characters. If used this
will allow test users to easily identify tests while still
leveraging the random_name in the create server behaviors.

Change-Id: If4305748e7e73bf5d8957c0f2971ef70fe6363de