Added name_prefix to create server behaviors in compute
Added name_prefix to create_server_with_defaults and
create_active_server. This will allow test users to specify a
name prefix for creating a server. This will be used with
random_name to create a server with a name that contains
the name_prefix value followed by random characters. If used this
will allow test users to easily identify tests while still
leveraging the random_name in the create server behaviors.

- Added name_prefix keyword argument to create server behaviors
- Added corresponding doc strings in the style currently found 
  for the name_prefix argument
- Modified arguments to create_server behavior to use keywords
  rather than being used as positional arguments

Change-Id: If4305748e7e73bf5d8957c0f2971ef70fe6363de