Added correct response model to get_limits method in limits_api
Previously the get_limits method in the limits_api client was
attempting to deserialize the limit for a get limits request
using the Limits model. However this was unsuccessful in
deserializing the body of the get limits request. Using the
TenantLimits response model this method is now capable of
successfully deserializing get limit responses.

- Changed import from Limits to TenantLimits
- Changed response type from Limits to TenantLimits in the
  get_limits method
- Updated copyright date
- Updated docstring to reflect changes

Change-Id: I818e9796444021f8ae4928d361baf1c7fb505d00