Modified _parse_resp_body in Compute Exception Handler
Modified _parse_resp_body in Compute Exception Handler in order
to allow correct removal of Namespaces in XML bodies. Additionally
removed 'finally' clause which would run regardless of the
try/except behaviors. The purpose of the original 'finally' clause
was to attempt to catch anything that could not be deserialized into
XML or JSON.

- Updated XML namespace removal in _parse_resp_body to point to
  the correct function
- Removed 'finally clause' to force a more intuitive code flow.

Change-Id: I7b94064e7c459a961b9063cd0c327b7bb02cc1a1